Lock to light mode, refactor files
