+++ export_logic/__init__.py
@@ -1,7 +1,4 @@
-"""
-This module contains the functionality for adding points to the session state
-and creating a GeoJSON FeatureCollection.
-"""
+"""Vector export logic: download filenames and format encoding."""
 
 import io
 import json
@@ -12,122 +9,6 @@ from pathlib import Path
 
 import geopandas as gpd
 import streamlit as st
-from shapely.geometry import Point
-
-
-def add_point(lat, lon, name_or_properties=""):
-    """Add a point to the session state.
-
-    Parameters
-    ----------
-    lat : float
-        The latitude coordinate of the point.
-    lon : float
-        The longitude coordinate of the point.
-    name_or_properties : str or dict, optional
-        Either a string name for the point or a dictionary of properties.
-        Defaults to empty string.
-
-    Returns
-    -------
-    None
-        Adds the point to st.session_state.points.
-    """
-    # Handle both string names and dictionary properties
-    try:
-        # Try to use as dictionary first
-        properties = name_or_properties.copy()
-        properties["timestamp"] = datetime.now().isoformat()
-        # Add a name if not present
-        if "name" not in properties:
-            properties["name"] = f"Point {len(st.session_state.points) + 1}"
-    except AttributeError:
-        # Not a dict, treat as string
-        properties = {
-            "name": (
-                name_or_properties
-                if name_or_properties
-                else f"Point {len(st.session_state.points) + 1}"
-            ),
-            "description": "",
-            "timestamp": datetime.now().isoformat(),
-        }
-
-    point = {
-        "type": "Feature",
-        "geometry": {"type": "Point", "coordinates": [lon, lat]},
-        "properties": properties,
-    }
-    st.session_state.points.append(point)
-
-
-def create_geojson():
-    """Create a GeoJSON FeatureCollection from stored points.
-
-    Returns
-    -------
-    dict
-        A GeoJSON FeatureCollection containing all stored points.
-    """
-    return {"type": "FeatureCollection", "features": st.session_state.points}
-
-
-def reset_points():
-    """Clear all points from the session state.
-
-    Returns
-    -------
-    None
-        Clears st.session_state.points and resets last_click.
-    """
-    st.session_state.points = []
-    st.session_state.last_click = None
-
-
-def points_to_gdf(points: list[dict]) -> gpd.GeoDataFrame:
-    """Convert session points (GeoJSON features) to a GeoDataFrame.
-
-    Parameters
-    ----------
-    points : list of dict
-        List of GeoJSON feature dictionaries.
-
-    Returns
-    -------
-    geopandas.GeoDataFrame
-        A GeoDataFrame containing the points with their properties.
-    """
-    if not points:
-        return gpd.GeoDataFrame()
-
-    geometries = []
-    properties_list = []
-
-    for pt in points:
-        lon, lat = pt["geometry"]["coordinates"]
-        geometries.append(Point(lon, lat))
-        properties = {
-            key: value
-            for key, value in pt["properties"].items()
-            if key not in ("lat", "lon")
-        }
-        properties_list.append({"lat": lat, "lon": lon, **properties})
-
-    gdf = gpd.GeoDataFrame(properties_list, geometry=geometries, crs="EPSG:4326")
-
-    return gdf
-
-
-def get_base_filename():
-    """Generate base filename with timestamp.
-
-    Returns
-    -------
-    str
-        A filename with format 'click2vector_YYYYMMDD_HHMMSS'.
-    """
-    return f"click2vector_{datetime.now().strftime('%Y%m%d_%H%M%S')}"
-
 
 EXPORT_EXTENSIONS: dict[str, str] = {
     "GeoJSON": ".geojson",
@@ -137,6 +18,17 @@ EXPORT_EXTENSIONS: dict[str, str] = {
 }
 
 _KNOWN_EXPORT_SUFFIXES = frozenset({".geojson", ".json", ".zip", ".fgb"})
+
+
+def get_base_filename() -> str:
+    """Generate base filename with timestamp.
+
+    Returns
+    -------
+    str
+        A filename with format 'click2vector_YYYYMMDD_HHMMSS'.
+    """
+    return f"click2vector_{datetime.now().strftime('%Y%m%d_%H%M%S')}"
 
 
 def build_export_filename(filename: str, export_type: str) -> str:
@@ -234,7 +126,7 @@ def _export_geojson_display(gdf: gpd.GeoDataFrame) -> str:
     return json.dumps(geojson_data, indent=2)
 
 
-def export_data(gdf, export_type):
+def export_data(gdf: gpd.GeoDataFrame, export_type: str) -> str | bytes:
     """Export GeoDataFrame to the specified format.
 
     Parameters
@@ -249,7 +141,6 @@ def export_data(gdf, export_type):
     str or bytes
         The exported data as a string (GeoJSON) or bytes (Shapefile/FlatGeobuf).
     """
-    # Use a mapping instead of if/elif chains
     export_functions = {
         "GeoJSON": _export_geojson_display,
         "Esri Shapefile (.zip)": _export_shapefile,
@@ -259,5 +150,4 @@ def export_data(gdf, export_type):
     try:
         return export_functions[export_type](gdf)
     except KeyError:
-        # Unknown export type
         return b""
