+++ import_logic/google_sheets.py
@@ -7,7 +7,7 @@ from urllib.parse import parse_qs, urlparse
 import pandas as pd
 import streamlit as st
 
-from click_to_geojson_functionality import add_point
+from locations import add_point
 
 
 def extract_sheet_id(sheets_url):
